extract `Profile` sub-component into `Header` division
